AutonomesFahren/requirements.txt
keras==2.15.0
tensorflow==2.15.0
pillow
numpy
customtkinter
matplotlib
opencv-python